--- src/main/java/gift/dto/MemberDTO.java
@@ -1,4 +1,0 @@
-package gift.dto;
-
-public record MemberDTO(String email, String password) {
-}
