--- src/main/java/gift/dto/WishDTO.java
@@ -1,4 +1,0 @@
-package gift.dto;
-
-public record WishDTO(Long memberId, Long productId) {
-}
