+++ src/main/java/gift/dto/request/CategoryRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.request;
 
 import jakarta.validation.constraints.NotNull;
