+++ src/main/java/gift/dto/request/MemberRequestDTO.java
@@ -1,0 +1,4 @@
+package gift.dto.request;
+
+public record MemberRequestDTO(String email, String password) {
+}
