+++ src/main/java/gift/dto/request/ProductRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.request;
 
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.Pattern;
