+++ src/main/java/gift/dto/request/TokenLoginRequestDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.request;
 
 public class TokenLoginRequestDTO {
     private String email;
