+++ src/main/java/gift/dto/response/ErrorResponse.java
@@ -1,3 +1,3 @@
-package gift.dto;
+package gift.dto.response;
 
 public record ErrorResponse(int statusCode, String message){}
