+++ src/main/java/gift/dto/response/ProductResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.response;
 
 public record ProductResponseDTO(Long id,
                                  String name,
