+++ src/main/java/gift/dto/response/WishResponseDTO.java
@@ -1,4 +1,4 @@
-package gift.dto;
+package gift.dto.response;
 
 public class WishResponseDTO{
